+++ source/compiler/sc6.c
@@ -755,7 +755,7 @@ SC_FUNC int assemble(FILE *fout,FILE *fin)
         assert(strlen(sym->name)<=sNAMEMAX);
         strcpy(alias,sym->name);
       } /* if */
-      nametablesize+=strlen(alias)+1;
+      nametablesize+=(int)strlen(alias)+1;
     } /* if */
   } /* for */
   assert(numnatives==ntv_funcid);
@@ -767,7 +767,7 @@ SC_FUNC int assemble(FILE *fout,FILE *fin)
       if (constptr->value>0) {
         assert(strlen(constptr->name)>0);
         numlibraries++;
-        nametablesize+=strlen(constptr->name)+1;
+        nametablesize+=(int)strlen(constptr->name)+1;
       } /* if */
     } /* for */
   } /* if */
@@ -778,7 +778,7 @@ SC_FUNC int assemble(FILE *fout,FILE *fin)
     if ((constptr->value & PUBLICTAG)!=0) {
       assert(strlen(constptr->name)>0);
       numtags++;
-      nametablesize+=strlen(constptr->name)+1;
+      nametablesize+=(int)strlen(constptr->name)+1;
     } /* if */
   } /* for */
 
@@ -841,8 +841,8 @@ SC_FUNC int assemble(FILE *fout,FILE *fin)
       pc_resetbin(fout,hdr.publics+count*sizeof(AMX_FUNCSTUBNT));
       pc_writebin(fout,&func,sizeof func);
       pc_resetbin(fout,nameofs);
-      pc_writebin(fout,sym->name,strlen(sym->name)+1);
-      nameofs+=strlen(sym->name)+1;
+      pc_writebin(fout,sym->name,(int)strlen(sym->name)+1);
+      nameofs+=(int)strlen(sym->name)+1;
       count++;
     } /* if */
   } /* for */
@@ -890,8 +890,8 @@ SC_FUNC int assemble(FILE *fout,FILE *fin)
       pc_resetbin(fout,hdr.natives+count*sizeof(AMX_FUNCSTUBNT));
       pc_writebin(fout,&func,sizeof func);
       pc_resetbin(fout,nameofs);
-      pc_writebin(fout,alias,strlen(alias)+1);
-      nameofs+=strlen(alias)+1;
+      pc_writebin(fout,alias,(int)strlen(alias)+1);
+      nameofs+=(int)strlen(alias)+1;
       count++;
     } /* for */
     free(nativelist);
@@ -912,8 +912,8 @@ SC_FUNC int assemble(FILE *fout,FILE *fin)
         pc_resetbin(fout,hdr.libraries+count*sizeof(AMX_FUNCSTUBNT));
         pc_writebin(fout,&func,sizeof func);
         pc_resetbin(fout,nameofs);
-        pc_writebin(fout,constptr->name,strlen(constptr->name)+1);
-        nameofs+=strlen(constptr->name)+1;
+        pc_writebin(fout,constptr->name,(int)strlen(constptr->name)+1);
+        nameofs+=(int)strlen(constptr->name)+1;
         count++;
       } /* if */
     } /* for */
@@ -934,8 +934,8 @@ SC_FUNC int assemble(FILE *fout,FILE *fin)
       pc_resetbin(fout,hdr.pubvars+count*sizeof(AMX_FUNCSTUBNT));
       pc_writebin(fout,&func,sizeof func);
       pc_resetbin(fout,nameofs);
-      pc_writebin(fout,sym->name,strlen(sym->name)+1);
-      nameofs+=strlen(sym->name)+1;
+      pc_writebin(fout,sym->name,(int)strlen(sym->name)+1);
+      nameofs+=(int)strlen(sym->name)+1;
       count++;
     } /* if */
   } /* for */
@@ -954,8 +954,8 @@ SC_FUNC int assemble(FILE *fout,FILE *fin)
       pc_resetbin(fout,hdr.tags+count*sizeof(AMX_FUNCSTUBNT));
       pc_writebin(fout,&func,sizeof func);
       pc_resetbin(fout,nameofs);
-      pc_writebin(fout,constptr->name,strlen(constptr->name)+1);
-      nameofs+=strlen(constptr->name)+1;
+      pc_writebin(fout,constptr->name,(int)strlen(constptr->name)+1);
+      nameofs+=(int)strlen(constptr->name)+1;
       count++;
     } /* if */
   } /* for */
@@ -1119,7 +1119,7 @@ static void append_dbginfo(FILE *fout)
         if (prevstr!=NULL) {
           assert(prevname!=NULL);
           dbghdr.files++;
-          dbghdr.size+=sizeof(cell)+strlen(prevname)+1;
+          dbghdr.size+=(int)(sizeof(cell)+strlen(prevname)+1);
         } /* if */
         previdx=codeidx;
       } /* if */
@@ -1130,7 +1130,7 @@ static void append_dbginfo(FILE *fout)
   if (prevstr!=NULL) {
     assert(prevname!=NULL);
     dbghdr.files++;
-    dbghdr.size+=sizeof(cell)+strlen(prevname)+1;
+    dbghdr.size+=(int)(sizeof(cell)+strlen(prevname)+1);
   } /* if */
 
   /* line number table */
@@ -1151,7 +1151,7 @@ static void append_dbginfo(FILE *fout)
       dbghdr.symbols++;
       name=strchr(str+2,':');
       assert(name!=NULL);
-      dbghdr.size+=sizeof(AMX_DBG_SYMBOL)+strlen(skipwhitespace(name+1));
+      dbghdr.size+=(int)(sizeof(AMX_DBG_SYMBOL)+strlen(skipwhitespace(name+1)));
       if ((prevstr=strchr(name,'['))!=NULL)
         while ((prevstr=strchr(prevstr+1,':'))!=NULL)
           dbghdr.size+=sizeof(AMX_DBG_SYMDIM);
@@ -1162,21 +1162,21 @@ static void append_dbginfo(FILE *fout)
   for (constptr=tagname_tab.next; constptr!=NULL; constptr=constptr->next) {
     assert(strlen(constptr->name)>0);
     dbghdr.tags++;
-    dbghdr.size+=sizeof(AMX_DBG_TAG)+strlen(constptr->name);
+    dbghdr.size+=(int)(sizeof(AMX_DBG_TAG)+strlen(constptr->name));
   } /* for */
 
   /* automaton table */
   for (constptr=sc_automaton_tab.next; constptr!=NULL; constptr=constptr->next) {
     assert(constptr->index==0 && strlen(constptr->name)==0 || strlen(constptr->name)>0);
     dbghdr.automatons++;
-    dbghdr.size+=sizeof(AMX_DBG_MACHINE)+strlen(constptr->name);
+    dbghdr.size+=(int)(sizeof(AMX_DBG_MACHINE)+strlen(constptr->name));
   } /* for */
 
   /* state table */
   for (constptr=sc_state_tab.next; constptr!=NULL; constptr=constptr->next) {
     assert(strlen(constptr->name)>0);
     dbghdr.states++;
-    dbghdr.size+=sizeof(AMX_DBG_STATE)+strlen(constptr->name);
+    dbghdr.size+=(int)(sizeof(AMX_DBG_STATE)+strlen(constptr->name));
   } /* for */
 
 
@@ -1210,7 +1210,7 @@ static void append_dbginfo(FILE *fout)
             aligncell(&previdx);
           #endif
           writeerror |= !pc_writebin(fout,&previdx,sizeof previdx);
-          writeerror |= !pc_writebin(fout,prevname,strlen(prevname)+1);
+          writeerror |= !pc_writebin(fout,prevname,(int)strlen(prevname)+1);
         } /* if */
         previdx=codeidx;
       } /* if */
@@ -1224,7 +1224,7 @@ static void append_dbginfo(FILE *fout)
       aligncell(&previdx);
     #endif
     writeerror |= !pc_writebin(fout,&previdx,sizeof previdx);
-    writeerror |= !pc_writebin(fout,prevname,strlen(prevname)+1);
+    writeerror |= !pc_writebin(fout,prevname,(int)strlen(prevname)+1);
   } /* if */
 
   /* line number table */
@@ -1280,7 +1280,7 @@ static void append_dbginfo(FILE *fout)
         align16(&dbgsym.dim);
       #endif
       writeerror |= !pc_writebin(fout,&dbgsym,offsetof(AMX_DBG_SYMBOL, name));
-      writeerror |= !pc_writebin(fout,symname,strlen(symname)+1);
+      writeerror |= !pc_writebin(fout,symname,(int)strlen(symname)+1);
       for (dim=0; dim<dbgsymdim; dim++) {
         #if BYTE_ORDER==BIG_ENDIAN
           align16(&dbgidxtag[dim].tag);
@@ -1299,7 +1299,7 @@ static void append_dbginfo(FILE *fout)
       align16(&id1);
     #endif
     writeerror |= !pc_writebin(fout,&id1,sizeof id1);
-    writeerror |= !pc_writebin(fout,constptr->name,strlen(constptr->name)+1);
+    writeerror |= !pc_writebin(fout,constptr->name,(int)strlen(constptr->name)+1);
   } /* for */
 
   /* automaton table */
@@ -1313,7 +1313,7 @@ static void append_dbginfo(FILE *fout)
     #endif
     writeerror |= !pc_writebin(fout,&id1,sizeof id1);
     writeerror |= !pc_writebin(fout,&address,sizeof address);
-    writeerror |= !pc_writebin(fout,constptr->name,strlen(constptr->name)+1);
+    writeerror |= !pc_writebin(fout,constptr->name,(int)strlen(constptr->name)+1);
   } /* for */
 
   /* state table */
@@ -1328,7 +1328,7 @@ static void append_dbginfo(FILE *fout)
     #endif
     writeerror |= !pc_writebin(fout,&id1,sizeof id1);
     writeerror |= !pc_writebin(fout,&id2,sizeof id2);
-    writeerror |= !pc_writebin(fout,constptr->name,strlen(constptr->name)+1);
+    writeerror |= !pc_writebin(fout,constptr->name,(int)strlen(constptr->name)+1);
   } /* for */
 
   delete_dbgstringtable();
